# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.CountAnomaliesByDeffectCategoriesResponseDTO:
    properties:
      count:
        type: integer
      deffect_categories:
        items:
          $ref: '#/definitions/entity.AnomaliesByDeffectCategory'
        type: array
      error:
        type: string
      success:
        type: boolean
    type: object
  dto.CountAnomaliesByOwnerCompaniesResponseDTO:
    properties:
      count:
        type: integer
      error:
        type: string
      owner_companies:
        items:
          $ref: '#/definitions/entity.AnomaliesByOwnerCompany'
        type: array
      success:
        type: boolean
    type: object
  dto.CountAnomaliesByServingCompaniesResponseDTO:
    properties:
      count:
        type: integer
      error:
        type: string
      serving_companies:
        items:
          $ref: '#/definitions/entity.AnomaliesByServingCompany'
        type: array
      success:
        type: boolean
    type: object
  dto.CountAnomaliesDTO:
    properties:
      count:
        type: integer
      percent:
        type: number
    type: object
  dto.CountAnomaliesGrouppedResponseDTO:
    properties:
      error:
        type: string
      groups_counts:
        items:
          $ref: '#/definitions/entity.CountAnomaliesGroupped'
        type: array
      success:
        type: boolean
    type: object
  dto.CountAnomaliesResponseDTO:
    properties:
      current:
        $ref: '#/definitions/dto.CountAnomaliesDTO'
      error:
        type: string
      previous:
        $ref: '#/definitions/dto.CountAnomaliesDTO'
      success:
        type: boolean
    type: object
  dto.CountPointsResponseDTO:
    properties:
      count:
        type: integer
      error:
        type: string
      success:
        type: boolean
    type: object
...
  entity.AnomaliesByDeffectCategory:
    properties:
      count:
        type: integer
      deffect_category:
        type: string
      percent:
        type: number
    type: object
  entity.AnomaliesByOwnerCompany:
    properties:
      count:
        type: integer
      owner_company:
        type: string
      percent:
        type: number
    type: object
  entity.AnomaliesByServingCompany:
    properties:
      count:
        type: integer
      percent:
        type: number
      serving_company:
        type: string
    type: object
...
  entity.CountAnomaliesGroupped:
    properties:
      count:
        type: integer
      percent:
        type: number
      type:
        type: integer
    type: object
...
host: ""
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: TechPotion ЛЦТ2022 Swagger
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger for TechPotion's leadershack2022 solution
  version: "1.0"
paths:
...
  /dashboard/anomalies/count:
    get:
      consumes:
      - application/json
      description: returns amount and percent of anomalies
      parameters:
      - description: Region
        in: query
        name: region
        required: true
        type: string
      - description: Lower time limit (RFC3339 formatted)
        in: query
        name: datetime_from
        required: true
        type: string
      - description: Upper time limit (RFC3339 formatted)
        in: query
        name: datetime_to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CountAnomaliesResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.CountAnomaliesResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.CountAnomaliesResponseDTO'
      summary: Count and get percent of anomalies
      tags:
      - dashboard
  /dashboard/anomalies/count_groupped:
    get:
      consumes:
      - application/json
      description: returns amount of anomalies groupped by cases
      parameters:
      - description: Region
        in: query
        name: region
        required: true
        type: string
      - description: Lower time limit (RFC3339 formatted)
        in: query
        name: datetime_from
        required: true
        type: string
      - description: Upper time limit (RFC3339 formatted)
        in: query
        name: datetime_to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CountAnomaliesGrouppedResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.CountAnomaliesGrouppedResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.CountAnomaliesGrouppedResponseDTO'
      summary: Count and get amount anomalies groupped by cases
      tags:
      - dashboard
  /dashboard/anomalies/ratings/deffect_categories:
    get:
      consumes:
      - application/json
      description: returns rating of anomalies groupped by serving companies based
        on deffect categories
      parameters:
      - description: Region
        in: query
        name: region
        required: true
        type: string
      - description: Lower time limit (RFC3339 formatted)
        in: query
        name: datetime_from
        required: true
        type: string
      - description: Upper time limit (RFC3339 formatted)
        in: query
        name: datetime_to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CountAnomaliesByDeffectCategoriesResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.CountAnomaliesByDeffectCategoriesResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.CountAnomaliesByDeffectCategoriesResponseDTO'
      summary: Get rating of anomalies groupped by deffect categories
      tags:
      - dashboard
  /dashboard/anomalies/ratings/owner_companies:
    get:
      consumes:
      - application/json
      description: returns rating of anomalies groupped by owner companies based on
        anomalies count
      parameters:
      - description: Region
        in: query
        name: region
        required: true
        type: string
      - description: Lower time limit (RFC3339 formatted)
        in: query
        name: datetime_from
        required: true
        type: string
      - description: Upper time limit (RFC3339 formatted)
        in: query
        name: datetime_to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CountAnomaliesByOwnerCompaniesResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.CountAnomaliesByOwnerCompaniesResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.CountAnomaliesByOwnerCompaniesResponseDTO'
      summary: Get rating of anomalies groupped by owner companies
      tags:
      - dashboard
  /dashboard/anomalies/ratings/serving_companies:
    get:
      consumes:
      - application/json
      description: returns rating of anomalies groupped by serving companies based
        on anomalies count
      parameters:
      - description: Region
        in: query
        name: region
        required: true
        type: string
      - description: Lower time limit (RFC3339 formatted)
        in: query
        name: datetime_from
        required: true
        type: string
      - description: Upper time limit (RFC3339 formatted)
        in: query
        name: datetime_to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CountAnomaliesByServingCompaniesResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.CountAnomaliesByServingCompaniesResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.CountAnomaliesByServingCompaniesResponseDTO'
      summary: Get rating of anomalies groupped by serving companies
      tags:
      - dashboard
...
  /points/count:
    get:
      consumes:
      - application/json
      description: returns amount of entities
      parameters:
      - description: Region
        in: query
        name: region
        required: true
        type: string
      - description: Lower time limit (RFC3339 formatted)
        in: query
        name: datetime_from
        required: true
        type: string
      - description: Upper time limit (RFC3339 formatted)
        in: query
        name: datetime_to
        required: true
        type: string
      - description: Bottom left X of screen
        in: query
        name: x_min
        type: number
      - description: Bottom left Y of screen
        in: query
        name: y_min
        type: number
      - description: Top right X of screen
        in: query
        name: x_max
        type: number
      - description: Top right Y of screen
        in: query
        name: y_max
        type: number
      - description: Serving companies (компания исполнитель)
        in: query
        items:
          type: string
        name: serving_company
        type: array
      - description: Efficiency (результативность)
        in: query
        items:
          type: string
        name: efficiency
        type: array
      - description: Grade for service (оценка качества)
        in: query
        items:
          type: string
        name: grade_for_service
        type: array
      - description: Urgency category (срочность)
        in: query
        items:
          type: string
        name: urgency_category
        type: array
      - description: Work type (вид работы)
        in: query
        items:
          type: string
        name: work_type
        type: array
      - description: Deffect category (категория дефекта)
        in: query
        items:
          type: string
        name: deffect_category
        type: array
      - description: Owner company (упр. организация)
        in: query
        items:
          type: string
        name: owner_company
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CountPointsResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.CountPointsResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.CountPointsResponseDTO'
      summary: Count points by selected area and region
      tags:
      - points
...
schemes:
- http
swagger: "2.0"